# specs/1-phase-2-part-1-task-crud-api/contracts/task-api-openapi.yaml
openapi: 3.0.3
info:
  title: Task Management API
  description: API for managing user tasks with CRUD operations
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
paths:
  /api/{user_id}/tasks:
    get:
      summary: Get all tasks for a user
      parameters:
        - name: user_id
          in: path
          required: true
          description: ID of the user
          schema:
            type: integer
      responses:
        '200':
          description: List of tasks for the user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '404':
          description: User not found
    post:
      summary: Create a new task for a user
      parameters:
        - name: user_id
          in: path
          required: true
          description: ID of the user
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: User not found
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationError'
  /api/{user_id}/tasks/{id}:
    get:
      summary: Get a specific task
      parameters:
        - name: user_id
          in: path
          required: true
          description: ID of the user
          schema:
            type: integer
        - name: id
          in: path
          required: true
          description: ID of the task
          schema:
            type: integer
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found or does not belong to user
    put:
      summary: Update a specific task
      parameters:
        - name: user_id
          in: path
          required: true
          description: ID of the user
          schema:
            type: integer
        - name: id
          in: path
          required: true
          description: ID of the task
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found or does not belong to user
        '422':
          description: Validation error
    delete:
      summary: Delete a specific task
      parameters:
        - name: user_id
          in: path
          required: true
          description: ID of the user
          schema:
            type: integer
        - name: id
          in: path
          required: true
          description: ID of the task
          schema:
            type: integer
      responses:
        '204':
          description: Task deleted successfully
        '404':
          description: Task not found or does not belong to user
  /api/{user_id}/tasks/{id}/complete:
    patch:
      summary: Update task completion status
      parameters:
        - name: user_id
          in: path
          required: true
          description: ID of the user
          schema:
            type: integer
        - name: id
          in: path
          required: true
          description: ID of the task
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                completed:
                  type: boolean
                  description: New completion status for the task
              required:
                - completed
      responses:
        '200':
          description: Task completion status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Task not found or does not belong to user
        '422':
          description: Validation error
components:
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the task
          example: 1
        title:
          type: string
          description: Title of the task
          example: "Buy groceries"
        description:
          type: string
          description: Detailed description of the task
          example: "Milk, bread, eggs"
        completed:
          type: boolean
          description: Completion status of the task
          example: false
        user_id:
          type: integer
          description: ID of the user who owns this task
          example: 1
        created_at:
          type: string
          format: date-time
          description: Timestamp when the task was created
          example: "2023-01-01T12:00:00Z"
        updated_at:
          type: string
          format: date-time
          description: Timestamp when the task was last updated
          example: "2023-01-01T12:00:00Z"
      required:
        - id
        - title
        - completed
        - user_id
        - created_at
        - updated_at
    TaskCreate:
      type: object
      properties:
        title:
          type: string
          description: Title of the task (1-255 characters)
          example: "Buy groceries"
          minLength: 1
          maxLength: 255
        description:
          type: string
          description: Detailed description of the task (optional)
          example: "Milk, bread, eggs"
          maxLength: 1000
        completed:
          type: boolean
          description: Initial completion status of the task
          example: false
      required:
        - title
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          description: Title of the task (1-255 characters)
          example: "Buy groceries"
          minLength: 1
          maxLength: 255
        description:
          type: string
          description: Detailed description of the task (optional)
          example: "Milk, bread, eggs"
          maxLength: 1000
        completed:
          type: boolean
          description: New completion status of the task
          example: true
      required:
        - title
    ValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
                description: Location of the validation error
              msg:
                type: string
                description: Error message
              type:
                type: string
                description: Type of validation error